* Added a property to TfpgCanvasImpl(x11) FastDoubleBuffer * Fixed painting of ListView Column * Misc Listview painting changes * Added OnColumnPaint to ListView
The new property FastDoubleBuffer will probably be moved to TfpgCanvaseBase also perhaps this can be set with an application 
property since it doesn't free the backbuffer until the Canvas is freed which will result in some increase in memory usage.

The listview can have double buffering disabled now and it won't flicker sonce the painting is done with no overlapping rects.