Fix Ctrl+RightArrow skipping a line
When you Ctrl+RightArrow to jump words, when it reached the end
of a line, it skipped a line and actually jumped two lines down.
Also when the caret landed in the new line it was at position 1
of the line and not at the true beginning of the line position 0.


Signed-off-by: Graeme Geldenhuys <graemeg@gmail.com>