docview: Give a more accurate draw width to TRichTextLayout.
This still contains one issue. We need to ask TRichTextLayout to
calculate lines and word wrapping, giving the maximum text width
available. Well, if text wrapping occurs and the text is long enough
to require scrolling, we suddenly need to take the scrollbar width
into account for word wrapping!! A chicken and egg situation. :-(
This needs to be fixed somehow!!

Signed-off-by: Graeme Geldenhuys <graemeg@gmail.com>