New "Close Button" parameter added to message dialog class functions.
if a Warning message dialog was shown with Yes/No buttons, and the user
canceled the dialog via Esc key or window frame X button, the dialog result
is mbCancel. The ACloseButton value can now replace the default mbCancel
result with whatever the user specifies.

If a mbCancel button is in the AButtons set, then ACloseButton is ignored.