Edit: fixed vertical centering of text.
Before we had a hard-coded top co-ordinate for the text. This
caused problems when the TfpgEdit.Height was adjusted for some
reason. So now we rather use the Canvas.DrawText() and give it
a rectangle and text flags so text output can be calculated
correctly. This is very useful for inline editing in a StringGrid.

Internally Canvas.DrawText calls fpgStyle.DrawString, so custom
styling should still be applied.