Implement OnShowHint for all relevant edit options.
The tooltips now show the values of each edit option, but as
expanded macros - the actual values that will be used.

We had to do some trickery with a composite components (editbuttons)
so we can access the correct component name to know what tooltip
values to use.

Also improved the label captions to better describe each option.