Catch the OS independent DROP[Enter|Exit] messages
Under X11 the XDND OS messages on only go to the top-level window, so we
send the FPGM_ messages ourselves. Later each component will hook into this
to change it's appearance to show the end-user that a drop is allowed or not.

eg: Some components will get a dark/black border or something.