Anchors now support relative positioning
If you only set a single anchor eg [anTop] and you resize the
width of the form, the component will stay relative to the
original position.

if you place a component in the center of a form, and set anchors
to [], and resize the form, the component will stay centered.