X11 DND: Incorrectly assumed there will always be a 'text/plain' mime-type
If the developer did not set the AMimeChoice in OnDragEnter, then
the incorrect default could have been used. By default it
used to deflaut to 'text/plain' and never actually checked the
mime-type list to see if that exists. :-(

Now the default AMimeChoice is set the the first mime-type in
 the mime-type list. This keeps with the methodology that the
mime-type list must be from most specific to least specific mime types.