X11 DND: We never checked if drop was really accepted in HandleDNDDrop
If the Drop occured, but AcceptDrops = False, we must not do
any Selections conversions. This is now so.

We must still send the XdndFinished message though, so data object
can be freed, and to simply comply with XDND protocol. We now
correctly set the Accept or Decine status in the XDNDFinished
message.