window grouping by wm
Attached (inline) is a patch for allowing multiple window groups.
Also, here is a link to the same, in case line-wrapping mangles it:
http://david9.freepgs.com/fpc/fpg_x11-multiple-window-groups.patch

I simply added fields for group leader as well as a special PXClassHint, and
used those in DoAllocateWindowHandle and SetWindowGroup.

In my own program, I manage this stuff by having a list of "res_class"es each
mapped to a group leader and PXClassHint. I override AfterCreate and fetch the
group leader and PXClassHint from that list, based on which res_class I'd like
the window to be grouped in.

It seems that the group leader part is not strictly necessary: if I leave it
unassigned and let the default FLeaderWindow be the group leader, the
PXClassHint assignment determines grouping (in KDE3 anyway. I haven't tested
other WM's)

So I'm not sure what the purpose of FLeaderWindow is at all, but I followed the
example and created the group leader.

~David.