new overloaded DumpStack() can now output to a TStringList
It wasn't always desireable for DumpStack to output to the console, because
it doesn't always exist, and the user will normally not see it if app is
launched from a icon.

We can now output that info to a stringlist, and display it to the end-user
in whatever way we want.