First draft of MDI support.
This work equates to about 3 hours of coding. Not bad I think! ;-)

The general idea is to have to new classes. The MDIWorkArea class is the
parent widget form MDI child forms. We then also have a MDIChildWindow
class which implements the outer child window - thus the one with the
titlebar and blue borders.

The end use creates a form based on TfpgFrame, not TfpgWindow. This design
might change later. The users form is then embedded inside the MDIChildWindow,
and displayed inside the bounds of the MDIWorkArea.