report: Fixes a cause of corrupt memory
The images use in a PDF report get freed later. But if we used internally registered image,
the report engine frees those images, even though they might still be needed elsewhere
in the application. So for that reason we "clone" the internal image, which the report
engine can safely free.