Fixes build error about units not for a specific target.
Because the X11 and GDI targets have units with the same names (ie: fpg_impl.pas etc)
the last enry overrides the platform target. So if we compile under Linux, it will complain that
fpg_impl.pas is for the Windows target only.

To work around this problem, we have to add some units in code blocks for specific targets.