richview: Removed the extra margin for text at the top and bottom of the client area
There was a hard-coded 2px margin at the top and bottom of the text client area. So when
scrolling the text disappeared 2px away from the actual widget border. This was inconsistent
with other text widgets in fpGUI, and I simply didn't like it either.

So GetTextAreaRect() now returns the same values as GetDrawRect().