Adds a new "Can Adjust Position" parameter to TfpgPopupWindow.ShowAt() call.
Sometimes we want to display a popup window, but would like the framework to take into
account the desktop screen size. If the popup window would display outside the bounds of
the screen, we would like fpGUI to reposition the popup window to be fully visible. This will
be used in the System Tray Icon implemenation.

This parameter defaults to False, so the behaviour is exactly like it was before this changes.