Implements WakeMainThread for Windows.
Multi-threaded apps under Windows requires the main thread to be woken up
periodically. We now simply hook into the Classes.WakeMainThread function
which is automatically called by TThread.Synchronize.

I did the same testing under X11, but it doesn't seem needed there. If we
do find a case for this under X11, it is easy to implement.