numeric edits: adds min/max limits
I was just ready to send a new patch concerning the limit values
which is solving the issue with the gridediting. In addition, it is now
possible to set only one limit, max or min, and I solved a bug on the
editfloat which prevented to enter decimals in some cases.