launching help viewer
Under *nix systems, if we can't find the direct location of docview, we
try the 'which docview' command. If we get a positive response, we know
docview is in the system path, and we can go ahead and launch it.

At least this doesn't instantly throw an exception when docview can't
be found on the first attempt.