Framework build scripts switched to DEBUG build mode
The reason this was done, is because the RELEASE mode allows FPC to strip
out code that isn't used by the framework. This causes a problem for other
projects, because other projects might still need that stripped out code
(eg: Docview). It run it often causes a "identifier not found" error.

Switching back to DEBUG build means those projects compile without problem.