ActiveTabColor/ActiveTabTextColor now defaults to clDefault
The problem was that if you didn't specify ActiveTab*Color values, but did
specify a Tabsheet to have a different color, when that tabsheet became
the active one the color changed to clWindowBackground. This looked odd.

Now the ActiveTab*Color = clDefault that tabsheet will still maintain its
color, even if it is a background tab or a active tab.

When you assign any other color value to ActiveTab*Color (not clDefault),
then they will take preference.

* Also prevent an error when calling DrawTab with the ATabSheet parameter
  being nil
* refactored some color assignment code