DND big changes.
X11 working, gdi not done yet.

TfpgDrag has an added param in the constructor. This is so dnd can be more efficient if the target and source are the same app instance. Drag is otherwise unchanged.

TfpgDrag.Create(ASource: TfpgWidgetBase);

OnDragEnter, OnDragLeave and OnDragDrop in TfpgWidget are gone. In their place is a DropHandler property. It can be assigned with
DropHandler := TfpgDropEventHandler.Create(@MyEnter, @MyLeave, @MyDrop @MyMove);
Create a subclass of TfpgDropHandler to implement a custom handler.

Added TfpgWidget.GetDefaultDropHandler: TfpgDropHandler; virtual;
This should be overridden by widgets that have builtin DND behavior like text widgets.

Added a FPGM_WINDOW_ALLOCATED message. This is triggered when a native window is allocated and set. The native window notifies it's primary widget that the window object is allocated and it has a valid WindowHandle. The primary widget then notifies it's children that don't have a native window. And so on.