x11: fixed a keyrelease event bug
The issue was raised by David Emerson. Using the eventtest application.
The following output was generated when Ctrl+S was pressed and released.

[] Key pressed: CtrlL
[Ctrl] Key pressed: S
[Ctrl] Key released: S
[Ctrl] Key released: S

Note the last line is a duplicate of line 3. This is the bug. The last
line should have been Ctrl was released.

With this patch the same key strokes now generate the following events:

[] Key pressed: CtrlL
[Ctrl] Key pressed: S
[Ctrl] Key released: S
[Ctrl] Key released: CtrlL