Fixes bug where File Dialog didn't display all files.
It didn't match the file extensions correctly and sometimes matched text
that was in the middle of the file name text, instead of the last file
extension.

I replaced all that processing with a simple regex expression. It could be
improved even more (eg: where multiple file extensions are specified). Those
could be treated as one regex expression, but I'll leave that for another
time.