TfpgMemo: consolidate & optimize X hit testing
Added TfpgMemo.CurPosByX() to determine where in a line the pixel
column X is. I've replaced the many occurrences of this algorithm
with a call to this function.

I made this function "private", like other related ones, but it and
many other methods should probably be upgraded to "protected" so that
a descendent class doesn't have to include duplicate code.