BUG: Fix TfpgSplitter.FindControl
It was unable to find controls that were closed all the way. This was
particularly bad with multiple controls stacked and aligned on a
side. You close one and then the splitter took over the other,
leaving the intended control lost.

Added components to splitter demo to test/prove the operation.