BUG: Fix failure to draw masked images in X11
On X11 TfpgCanvas.DrawImage() failed to draw an image when it was
masked after a call to TfpgCanvas.ClearClipRect(), which happens
when HandlePaint is called. This prevented things like message dialog
icons and radio buttons from showing.

Also removed extra Canvas.Clear(BackgroundColor) call when "about"
dialog is rendered in TfpgMessageDialog.HandlePain.